feat: added automatic user id joining from token

+++ API/Controllers/FollowersController.cs
@@ -1,4 +1,5 @@
-﻿using BLL.DTO.Follower;
+﻿using System.Security.Claims;
+using BLL.DTO.Follower;
 using BLL.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -20,6 +21,11 @@ namespace API.Controllers
         [HttpPost]
         public async Task<IActionResult> Follow([FromBody] FollowerDto dto, CancellationToken cancellationToken)
         {
+            var userIdFromToken = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userIdFromToken == null || !int.TryParse(userIdFromToken, out var userId))
+                return Unauthorized("Invalid token.");
+
+            dto.FollowerID = userId;
             var id = await _followerService.AddFollowerAsync(dto, cancellationToken);
             return Ok(new { id });
         }
@@ -28,19 +34,29 @@ namespace API.Controllers
         [HttpDelete]
         public async Task<IActionResult> Unfollow([FromBody] FollowerDto dto, CancellationToken cancellationToken)
         {
+            var userIdFromToken = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userIdFromToken == null || !int.TryParse(userIdFromToken, out var userId))
+                return Unauthorized("Invalid token.");
+
+            dto.FollowerID = userId;
             await _followerService.DeleteAsync(dto, cancellationToken);
             return NoContent();
         }
 
         [Authorize]
         [HttpGet("is-following")]
-        public async Task<IActionResult> IsFollowing([FromQuery] int followerId, [FromQuery] int followingId)
+        public async Task<IActionResult> IsFollowing([FromQuery] int followingId)
         {
+            var userIdFromToken = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            if (userIdFromToken == null || !int.TryParse(userIdFromToken, out var followerId))
+                return Unauthorized("Invalid token.");
+
             var result = await _followerService.IsFollowingAsync(new FollowerDto
             {
                 FollowerID = followerId,
                 FollowingID = followingId
             });
+
             return Ok(result);
         }
